# fix itoa bench: reuse the same buffer and avoid redundant allocation during each iter