fix https://github.com/ohkami-rs/ohkami/commit/8fcc4f67f888e6a3134fbaba03862e05d7e07695#r143970218

> 🤔
> 
> It seems to neglect the fact that the process specified in bench is executed repeatedly. The size of buf is increased each time the contents of bench are executed.

---

*benches/benches/itoa.rs*
```diff
let mut buf = String::with_capacity(v.len() * 21);
b.iter(|| {
+    /*
+        reuse the same buffer and avoid redundant allocation during each iter
+
+        - `.as_mut_vec().set_len(0)` is very cheap at the cost
+        - `buf.{push, push_str}` just do `ptr::write` and `set_len` due to the enough capacity,
+          whitch means following `c_lib` overwrites existing bytes
+        - `Drop` impls of overwritten bytes are never called but this is not a problem
+          because it (`u8::drop`) is noop
+    */
+    unsafe {buf.as_mut_vec().set_len(0)}
+
    c_lib(&mut buf)
});
```

---

## benchmark result samples

- before

```txt
running 21 tests
test a_small_http_response_content_length::itoa_01        ... bench:     179,831.11 ns/iter (+/- 3,578.60)
test a_small_http_response_content_length::itoa_02        ... bench:     155,546.78 ns/iter (+/- 3,622.06)
test a_small_http_response_content_length::itoa_03        ... bench:     179,853.00 ns/iter (+/- 4,012.93)
test a_small_http_response_content_length::itoa_05        ... bench:     130,003.99 ns/iter (+/- 28,542.64)
test a_small_http_response_content_length::itoa_06        ... bench:     146,642.52 ns/iter (+/- 39,976.65)
test a_small_http_response_content_length::itoa_07        ... bench:     145,467.02 ns/iter (+/- 6,858.02)
test a_small_http_response_content_length::itoa_to_string ... bench:     259,252.72 ns/iter (+/- 35,008.44)
test common_json_response_content_length::itoa_01         ... bench:     343,990.17 ns/iter (+/- 87,475.72)
test common_json_response_content_length::itoa_02         ... bench:     318,295.19 ns/iter (+/- 11,735.58)
test common_json_response_content_length::itoa_03         ... bench:     343,391.40 ns/iter (+/- 47,059.37)
test common_json_response_content_length::itoa_05         ... bench:     230,475.57 ns/iter (+/- 13,283.69)
test common_json_response_content_length::itoa_06         ... bench:     201,010.16 ns/iter (+/- 10,688.55)
test common_json_response_content_length::itoa_07         ... bench:     207,596.80 ns/iter (+/- 22,292.77)
test common_json_response_content_length::itoa_to_string  ... bench:     304,877.13 ns/iter (+/- 15,720.21)
test http_response_content_length::itoa_01                ... bench:     618,571.40 ns/iter (+/- 16,448.52)
test http_response_content_length::itoa_02                ... bench:     573,146.40 ns/iter (+/- 18,646.34)
test http_response_content_length::itoa_03                ... bench:     702,511.26 ns/iter (+/- 223,131.39)
test http_response_content_length::itoa_05                ... bench:     344,653.31 ns/iter (+/- 153,406.10)
test http_response_content_length::itoa_06                ... bench:     388,013.30 ns/iter (+/- 45,917.81)
test http_response_content_length::itoa_07                ... bench:     368,911.80 ns/iter (+/- 154,183.86)
test http_response_content_length::itoa_to_string         ... bench:     384,258.72 ns/iter (+/- 22,120.96)
```

- after

```txt
running 21 tests
test a_small_http_response_content_length::itoa_01        ... bench:     167,287.71 ns/iter (+/- 2,412.22)
test a_small_http_response_content_length::itoa_02        ... bench:     141,031.73 ns/iter (+/- 1,565.03)
test a_small_http_response_content_length::itoa_03        ... bench:     167,027.35 ns/iter (+/- 2,985.24)
test a_small_http_response_content_length::itoa_05        ... bench:     116,866.07 ns/iter (+/- 1,858.86)
test a_small_http_response_content_length::itoa_06        ... bench:     108,887.43 ns/iter (+/- 1,894.69)
test a_small_http_response_content_length::itoa_07        ... bench:     109,621.52 ns/iter (+/- 2,313.39)
test a_small_http_response_content_length::itoa_to_string ... bench:     197,070.68 ns/iter (+/- 8,730.06)
test common_json_response_content_length::itoa_01         ... bench:     284,067.63 ns/iter (+/- 6,314.98)
test common_json_response_content_length::itoa_02         ... bench:     245,518.90 ns/iter (+/- 3,104.71)
test common_json_response_content_length::itoa_03         ... bench:     245,180.92 ns/iter (+/- 11,700.00)
test common_json_response_content_length::itoa_05         ... bench:     164,656.15 ns/iter (+/- 5,491.40)
test common_json_response_content_length::itoa_06         ... bench:     137,898.09 ns/iter (+/- 3,833.79)
test common_json_response_content_length::itoa_07         ... bench:     139,787.11 ns/iter (+/- 6,362.44)
test common_json_response_content_length::itoa_to_string  ... bench:     217,264.94 ns/iter (+/- 5,317.81)
test http_response_content_length::itoa_01                ... bench:     435,857.01 ns/iter (+/- 9,996.90)
test http_response_content_length::itoa_02                ... bench:     406,646.65 ns/iter (+/- 2,415.08)
test http_response_content_length::itoa_03                ... bench:     371,811.25 ns/iter (+/- 10,533.79)
test http_response_content_length::itoa_05                ... bench:     198,992.07 ns/iter (+/- 7,535.14)
test http_response_content_length::itoa_06                ... bench:     181,097.04 ns/iter (+/- 6,196.07)
test http_response_content_length::itoa_07                ... bench:     179,543.22 ns/iter (+/- 2,831.80)
test http_response_content_length::itoa_to_string         ... bench:     285,790.62 ns/iter (+/- 53,679.98)
```